Use usize for addresses rather than fixed size u64

+++ src/hooks.rs
@@ -81,7 +81,12 @@ unsafe fn compare_mask(addr: *const u8, op_mask: &[u8], str_mask: &str) -> bool 
 /// * end - The address to end searching at
 /// * op_mask - The opcodes to match against
 /// * str_mask - The str pattern defining how to match the opcodes
-unsafe fn find_pattern(start: u64, end: u64, op_mask: &[u8], str_mask: &str) -> Option<*const u8> {
+unsafe fn find_pattern(
+    start: usize,
+    end: usize,
+    op_mask: &[u8],
+    str_mask: &str,
+) -> Option<*const u8> {
     (start..=end)
         .map(|addr| addr as *const u8)
         .find(|addr| compare_mask(*addr, op_mask, str_mask))
@@ -97,8 +102,8 @@ pub unsafe fn hook() {
 unsafe fn verify_certificate() {
     // Last known addr: 0x0000000140CDCB30
 
-    let start_addr: u64 = 0x0000000140100000;
-    let end_addr: u64 = 0x0000000160000000;
+    let start_addr: usize = 0x0000000140100000;
+    let end_addr: usize = 0x0000000160000000;
 
     // Find the pattern for VerifyCertificate
     let call_addr = find_pattern(
@@ -161,8 +166,8 @@ unsafe fn verify_certificate() {
 unsafe fn bcrypt_verify_signature() {
     // Last known addr: 0x0000000140CDCB30
 
-    let start_addr: u64 = 0x0000000140100000;
-    let end_addr: u64 = 0x0000000200000000;
+    let start_addr: usize = 0x0000000140100000;
+    let end_addr: usize = 0x0000000200000000;
 
     // Find the pattern for VerifyCertificate
     let call_addr = find_pattern(
